Updated to look for the number of visible devices when waiting for dask workers.
