+++ python/cugraph/tests/dask/test_mg_louvain.py
@@ -14,13 +14,11 @@
 import pytest
 
 import cugraph.dask as dcg
-import cugraph.comms as Comms
-from dask.distributed import Client
 import cugraph
 import dask_cudf
-from dask_cuda import LocalCUDACluster
 from cugraph.tests import utils
-from cugraph.dask.common.mg_utils import is_single_gpu
+from cugraph.dask.common.mg_utils import (is_single_gpu,
+                                          setup_local_dask_cluster)
 
 try:
     from rapids_pytest_benchmark import setFixtureParamNames
@@ -45,15 +43,12 @@ except ImportError:
 @pytest.fixture(scope="module")
 def client_connection():
     # setup
-    cluster = LocalCUDACluster()
-    client = Client(cluster)
-    client.wait_for_workers(None)  # number of devices None = all vsble devices
-    Comms.initialize(p2p=True)
+    (comms, client, cluster) = setup_local_dask_cluster(p2p=True)
 
     yield client
 
     # teardown
-    Comms.destroy()
+    comms.destroy()
     client.close()
     cluster.close()
